requirements.txt
streamlit
pandas
numpy
Pillow
pytesseract
spacy
torch
transformers
peft
accelerate
safetensors
huggingface-hub
tokenizers
